Moved AlbumShow to correct folder
