+++ ui/src/artist/index.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import ArtistList from './ArtistList'
+import ArtistShow from './ArtistShow'
 import DynamicMenuIcon from '../layout/DynamicMenuIcon'
 import MicNoneOutlinedIcon from '@material-ui/icons/MicNoneOutlined'
 import MicIcon from '@material-ui/icons/Mic'
-import ArtistShow from '../common/ArtistShow'
 
 export default {
   list: ArtistList,
